Fix some of the Windows errors

+++ schemes/mceliece460896f/vec/bm.c
@@ -14,7 +14,7 @@ static inline uint16_t mask_nonzero(gf a) {
     ret >>= 31;
     ret -= 1;
 
-    return ret;
+    return (uint16_t)ret;
 }
 
 static inline uint16_t mask_leq(uint16_t a, uint16_t b) {
@@ -25,7 +25,7 @@ static inline uint16_t mask_leq(uint16_t a, uint16_t b) {
     ret >>= 31;
     ret -= 1;
 
-    return ret;
+    return (uint16_t)ret;
 }
 
 static inline void vec_cmov(vec *out, const vec *in, uint16_t mask) {
